+++ src/grammar/util/string_functions.h
@@ -1,4 +1,4 @@
 #ifndef __STR_FNS_H__
 #define __STR_FNS_H__
-char * concat_strings(char ** s, unsigned long long num_s);
+char * concat_strings(char ** s, unsigned int num_s);
 #endif
